refactor because of cognitive complexity
