+++ src/components/RichTextEditor/serializer/md/RemarkPlate/deserialize.ts
@@ -1,3 +1,5 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
 import {
     BlockQuoteNode,
     CodeBlockNode,
