+++ src/components/RichTextEditor/serializer/md/RemarkPlate/index.ts
@@ -1,3 +1,5 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
 /**
  * The main source
  * https://github.com/hanford/remark-slate
