+++ src/components/RichTextEditor/serializer/md/RemarkPlate/serialize/index.ts
@@ -1,6 +1,10 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
 import { BlockType, InputNodeTypes, LeafType, defaultNodeTypes } from '../astTypes';
+import { applyFormattingToLeafNode } from './appplyFormatingToLeafNode';
 import { isLeafNode } from './isLeafNode';
 import { processNodes } from './processNodes';
+import { BREAK_TAG } from './utils';
 
 interface Options {
     nodeTypes?: InputNodeTypes;
@@ -9,8 +13,6 @@ interface Options {
 }
 
 const VOID_ELEMENTS: Array<keyof InputNodeTypes> = ['thematic_break', 'image'];
-
-const BREAK_TAG = '<br>';
 
 export default function serialize(chunk: BlockType | LeafType, opts: Options = { nodeTypes: defaultNodeTypes }) {
     const { nodeTypes: userNodeTypes = defaultNodeTypes, ignoreParagraphNewline = false, listDepth = 0 } = opts;
@@ -88,68 +90,7 @@ export default function serialize(chunk: BlockType | LeafType, opts: Options = {
         return;
     }
 
-    // Never allow decorating break tags with rich text formatting,
-    // this can malform generated markdown
-    // Also ensure we're only ever applying text formatting to leaf node
-    // level chunks, otherwise we can end up in a situation where
-    // we try applying formatting like to a node like this:
-    // "Text foo bar **baz**" resulting in "**Text foo bar **baz****"
-    // which is invalid markup and can mess everything up
-    if (children !== BREAK_TAG && isLeafNode(chunk)) {
-        if (chunk.strikeThrough && chunk.bold && chunk.italic) {
-            children = retainWhitespaceAndFormat(children, '~~***');
-        } else if (chunk.bold && chunk.italic) {
-            children = retainWhitespaceAndFormat(children, '***');
-        } else {
-            if (chunk.bold) {
-                children = retainWhitespaceAndFormat(children, '**');
-            }
-
-            if (chunk.italic) {
-                children = retainWhitespaceAndFormat(children, '_');
-            }
-
-            if (chunk.strikeThrough) {
-                children = retainWhitespaceAndFormat(children, '~~');
-            }
-
-            if (chunk.code) {
-                children = retainWhitespaceAndFormat(children, '`');
-            }
-        }
-    }
+    children = applyFormattingToLeafNode(children, chunk);
 
     return processNodes(type, nodeTypes, children, chunk, listDepth);
 }
-
-// This function handles the case of a string like this: "   foo   "
-// Where it would be invalid markdown to generate this: "**   foo   **"
-// We instead, want to trim the whitespace out, apply formatting, and then
-// bring the whitespace back. So our returned string looks like this: "   **foo**   "
-function retainWhitespaceAndFormat(string: string, format: string) {
-    // we keep this for a comparison later
-    const frozenString = string.trim();
-
-    // children will be mutated
-    const children = frozenString;
-
-    // We reverse the right side formatting, to properly handle bold/italic and strikeThrough
-    // formats, so we can create ~~***FooBar***~~
-    const fullFormat = `${format}${children}${reverseStr(format)}`;
-
-    // This conditions accounts for no whitespace in our string
-    // if we don't have any, we can return early.
-    if (children.length === string.length) {
-        return fullFormat;
-    }
-
-    // if we do have whitespace, let's add our formatting around our trimmed string
-    // We reverse the right side formatting, to properly handle bold/italic and strikeThrough
-    // formats, so we can create ~~***FooBar***~~
-    const formattedString = format + children + reverseStr(format);
-
-    // and replace the non-whitespace content of the string
-    return string.replace(frozenString, formattedString);
-}
-
-const reverseStr = (string: string) => string.split('').reverse().join('');
