+++ src/components/RichTextEditor/serializer/md/RemarkPlate/serialize/isLeafNode.ts
@@ -1,3 +1,5 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
 import { BlockType, LeafType } from '../astTypes';
 
 export const isLeafNode = (node: BlockType | LeafType): node is LeafType => {
