+++ src/components/RichTextEditor/serializer/md/RemarkPlate/serialize/utils.ts
@@ -1,0 +1,3 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+export const BREAK_TAG = '<br>';
